Create separate object that "ticks" game time
