+++ app/src/main/java/com/cannawen/dota2timer/game/interfaces/Game.java
@@ -1,8 +1,9 @@
 package com.cannawen.dota2timer.game.interfaces;
 
 import com.cannawen.dota2timer.configuration.Configuration;
+import com.cannawen.dota2timer.timer.TimerListener;
 
-public interface Game {
+public interface Game extends TimerListener {
     void start();
 
     void end();
