+++ app/src/main/java/com/cannawen/dota2timer/timer/TimerListener.java
@@ -1,0 +1,5 @@
+package com.cannawen.dota2timer.timer;
+
+public interface TimerListener {
+    void tick();
+}
